requirements.txt
fastapi
uvicorn
openai
pytesseract
pillow
sqlalchemy
PyPDF2
